% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_bivariee.R
\name{ancov_biv}
\alias{ancov_biv}
\title{regression with 2 independant variables}
\usage{
ancov_biv(model, showShapiro = T, showBartlett = T)
}
\arguments{
\item{model}{a model object}

\item{showShapiro}{if True, Shapiro test is realised on residuals to test for normality}

\item{showBartlett}{if True, Bartlett test is realised on residuals to test for equality of variance}
}
\value{
function returns a list of 3 object : "model" containing model summary, "shapiro pvalue" and "bartlett pvalue"
}
\description{
regression with 2 independant variables
}
