% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_multivariee.R
\name{ancov_multiv}
\alias{ancov_multiv}
\title{regression with multiple covariables}
\usage{
ancov_multiv(model, showCI = F)
}
\arguments{
\item{model}{a model object}

\item{showCI}{whether to show a second output with confidence intervals for coefficients}
}
\value{
function returns model summary and coefficients with confidence intervals
}
\description{
regression with multiple covariables
}
