% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_quanti.R
\name{describe_quanti}
\alias{describe_quanti}
\title{describe numerical variables of the provided dataset (mean, sd, min, max, median,
iqr, ci)}
\usage{
describe_quanti(data, createCSV = F)
}
\arguments{
\item{data}{a data set with only numerical variables}

\item{createCSV}{if True, a CSV file is created}
}
\value{
the function writes a csv, with variables as rows and mean, sd, min, max... as columns
}
\description{
describe numerical variables of the provided dataset (mean, sd, min, max, median,
iqr, ci)
}
