% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptionv0v1.R
\name{descri}
\alias{descri}
\title{describe}
\usage{
descri(vecteurj0, vecteurj90, groupe)
}
\arguments{
\item{vecteurj0}{a numerical vector (t0)}

\item{vecteurj90}{a numerical vector (t+1)}

\item{groupe}{a group variable}
}
\value{
return a vector of strings values, corresponding to N, mean, sd, se rounded up to 2 decimals
}
\description{
describe
}
